+++ tests/test_convert.py
@@ -55,6 +55,7 @@ tests = [
     "de_bw",
     "lt_kzs",
     "de_by_block",
+    "de_sl_block",
     "de_he",
     "de_st",
 ]
@@ -83,6 +84,7 @@ extra_convert_parameters = {
     "lt_kzs": _input_files("lt_kzs", "lt_kzs.json"),
     "de_by_block": _input_files("de_by_block", "de_by_block.gml"),
     "de_st": _input_files("de_st", "de_st.json"),
+    "de_sl_block": _input_files("de_sl_block", "de_sl_block.gml"),
 }